requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
boto3==1.29.7
django-cors-headers==4.3.1
django-filter==23.5
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0